Use aria-sort instead of a class
This makes the sort more accessible to JAWS users (sadly, VoiceOver
doesn't do anything with the information).

It also retains the information about which way a column was sorted.
Nothing tells the user which column they sorted last except memory, but
for setting up a primary, secondary, and tertiary sort that's typically
enough ...