Include td data-sort-column-key="price" in demo
The example markup for the "Sorting by column keys" demo includes the key selector in the header th cells, but not their td cells. Without this, the example is confusing as it is unclear how the selector maps to the cells. The interactive table in the demo is correct, so I suggest updating the example markup to match.